backend/requirements.txt
# ── Core Framework ─────────────────────────────────────────────
fastapi
uvicorn[standard]

# ── Google Agent Development Kit ────────────────────────────────
google-adk                                 # Agent orchestration & routing

# ── Google Cloud SDKs ───────────────────────────────────────────
google-cloud-aiplatform                    # Vertex AI: Gemini 2.5 Flash + Text Embeddings
google-api-python-client                   # Calendar, Drive, Gmail, Docs APIs
google-auth                                # Application Default Credentials
google-auth-oauthlib                       # OAuth2 for Google Workspace APIs
google-auth-httplib2                       # HTTP transport for Google auth
google-cloud-logging                       # Cloud Logging (agent trace streaming)
google-cloud-pubsub                        # Pub/Sub for event-driven triggers
google-cloud-secret-manager                # Secret Manager for API keys

# ── Database ────────────────────────────────────────────────────
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pgvector==0.2.5
alembic                                    # Database migrations

# ── Observability ───────────────────────────────────────────────
langwatch
openinference-instrumentation-google-adk

# ── Config & Utils ──────────────────────────────────────────────
pydantic-settings==2.2.1
python-dotenv
httpx
youtube-transcript-api                     # YouTube transcript data source
tenacity
